requirements.txt
asgiref==3.9.1
Django==4.2
djangorestframework==3.16.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
tzdata==2025.2
gunicorn
celery
redis
pandas
openpyxl